Typo, intervals --> intervales
Page scan shows "intervales." Editors probably thought they were correcting a typo in the original, but "intervales" is a real word that  means bottomland, or a low-lying land along a watercourse. The word is relevant and makes sense in the context here.

https://www.merriam-webster.com/dictionary/intervales

Also, most other editions show "intervales", although a couple of newish editions made the same change to "intervals"

https://archive.org/search?query=%28%22places+and+silent%22%29+AND+title%3A%28the+blue+castle%29+AND+mediatype%3A%28texts%29&sin=TXT&sort=-date